Split out chat panel changes — execution UI only
Remove TestFlowDialog, TestFlowButton container, chat avatars, react-markdown/remark-gfm/fetch-event-source deps, uuid/fileMock mocks, and related vite proxy changes. These belong to feat-agentflow-chatpanel which will be opened on top of this branch.

Restore StyledFab to node-palette, restore old TestFlowButton component, and revert Agentflow.tsx/canvas index to pre-chat state so this branch contains only the node execution status UI work.
