+++ packages/agentflow/examples/src/demos/E2eExample.tsx
@@ -338,7 +338,6 @@ export function E2eExample() {
                             apiBaseUrl={apiBaseUrl}
                             token={token ?? undefined}
                             initialFlow={loadedFlow}
-                            flowId={activeChatflowId}
                             onFlowChange={handleFlowChange}
                             onSave={handleSave}
                             canvasActions={canvasActions}
