+++ packages/agentflow/package.json
@@ -83,10 +83,7 @@
         "@uiw/codemirror-theme-sublime": "^4.21.0",
         "@uiw/codemirror-theme-vscode": "^4.21.0",
         "@uiw/react-codemirror": "^4.21.0",
-        "@microsoft/fetch-event-source": "^2.0.1",
         "axios": "1.15.0",
-        "react-markdown": "^9.1.0",
-        "remark-gfm": "^4.0.1",
         "dompurify": "^3.2.6",
         "flowise-react-json-view": "^1.21.7",
         "html-react-parser": "^3.0.16",
