--- packages/agentflow/src/__mocks__/react-markdown.tsx
@@ -1,4 +1,0 @@
-import React from 'react'
-
-const ReactMarkdown = ({ children }: { children?: React.ReactNode }) => <span>{children}</span>
-export default ReactMarkdown
