--- packages/agentflow/src/__mocks__/remark-gfm.ts
@@ -1,2 +1,0 @@
-const remarkGfm = () => {}
-export default remarkGfm
