--- packages/agentflow/src/__mocks__/uuid.ts
@@ -1,7 +1,0 @@
-let _counter = 0
-
-export const v4 = (): string => `test-uuid-${++_counter}`
-
-export const reset = (): void => {
-    _counter = 0
-}
