+++ packages/agentflow/src/atoms/index.ts
@@ -17,7 +17,6 @@ export { RichTextEditor } from './RichTextEditor.lazy'
 export { ScenariosInput, type ScenariosInputProps } from './ScenariosInput'
 export { StateKeyValueInput, type StateKeyValueInputProps } from './StateKeyValueInput'
 export { StructuredOutputBuilder, type StructuredOutputBuilderProps, type StructuredOutputEntry } from './StructuredOutputBuilder'
-export { StyledFab, type StyledFabProps } from './StyledFab'
 export { SuggestionDropdown, type SuggestionDropdownProps, type SuggestionDropdownRef, type SuggestionItem } from './SuggestionDropdown'
 export { SwitchInput, type SwitchInputProps } from './SwitchInput'
 export { TooltipWithParser, type TooltipWithParserProps } from './TooltipWithParser'
