+++ packages/agentflow/src/features/canvas/components/index.ts
@@ -7,4 +7,5 @@ export { NodeModelConfigs } from './NodeModelConfigs'
 export { getMinimumNodeHeight, NodeOutputHandles } from './NodeOutputHandles'
 export { NodeStatusIndicator, NodeWarningIndicator } from './NodeStatusIndicator'
 export { NodeToolbarActions } from './NodeToolbarActions'
+export { TestFlowButton } from './TestFlowButton'
 export { ValidationFeedback } from './ValidationFeedback'
