+++ packages/agentflow/src/features/canvas/containers/index.ts
@@ -3,7 +3,3 @@ export { AgentFlowNode } from './AgentFlowNode'
 export { IterationNode } from './IterationNode'
 export { NodeInfoDialog } from './NodeInfoDialog'
 export { StickyNote } from './StickyNote'
-export type { TestFlowButtonProps } from './TestFlowButton'
-export { TestFlowButton } from './TestFlowButton'
-export type { TestFlowDialogProps } from './TestFlowDialog'
-export { TestFlowDialog } from './TestFlowDialog'
