+++ packages/agentflow/src/features/node-palette/AddNodesDrawer.tsx
@@ -20,13 +20,14 @@ import {
 import { useTheme } from '@mui/material/styles'
 import { IconMinus, IconPlus, IconSearch, IconX } from '@tabler/icons-react'
 
-import { MainCard, StyledFab } from '@/atoms'
+import { MainCard } from '@/atoms'
 import { tokens } from '@/core/theme/tokens'
 import type { NodeDataSchema } from '@/core/types'
 import { useApiContext } from '@/infrastructure/store'
 
 import { NodeListItem } from './NodeListItem'
 import { debounce, groupNodesByCategory, searchNodes } from './search'
+import { StyledFab } from './StyledFab'
 import { useDrawerMaxHeight } from './useDrawerMaxHeight'
 
 export interface AddNodesDrawerProps {
