+++ packages/agentflow/src/features/node-palette/index.ts
@@ -1,3 +1,4 @@
 // Node Palette Feature - Public API
 export type { AddNodesDrawerProps } from './AddNodesDrawer'
 export { AddNodesDrawer } from './AddNodesDrawer'
+export { StyledFab } from './StyledFab'
